+++ changelog.md
@@ -1,3 +1,4 @@
 ## Bug fixes
+- toggling `enhancedTable` is possible through the `Legend API`
 
 ## New Features